refactor(studio-integration): Extract Observables Assertion
To streamline assertions for Observables, extracting method, which
evolved while testing several observables in studio-integration-mock.
